Feature [Notes]: Vista y validación de notas acorde a score type y creación de unidades
- Actualización de almacenamiento de calificaciones durante creación, revisión y edición de actividades para considerar el tipo de calificación.
- Incorporación de la capacidad para crear unidades al crear un curso, sin permitir la futura modificación de estas.
- Validaciones para verificar características de unidad y actividad para acoplarse al tipo de calificación.
- Creación de DTOs para crear y visualizar unidades.
- Actualización de servicios para visualizar notas tomando en consideración el nuevo almacenamiento de notas.
Fixed #152

+++ ClassNotes.API/ClassNotes.API/Services/DashboardCourses/DashboardCoursesService.cs
@@ -206,7 +206,8 @@ namespace ClassNotes.API.Services.DashboardCourses
                     // Calcular el ponderado de esta actividad
                     if (totalMaxScores > 0 && activity.Unit.MaxScore > 0)
                     {
-                        float weightedScore = (activity.MaxScore / totalMaxScores) * activity.Unit.MaxScore; // Aqui se hace la ponderación
+                        //(Ken) forzado a float para que no de errores, debera arreglarse para considerar los nulos de las unidades tipo oro...
+                        float weightedScore = (float)((activity.MaxScore / totalMaxScores) * activity.Unit.MaxScore); // Aqui se hace la ponderación
                         totalScoreEvaluated += weightedScore;
                     }
                 }
